lib/scrapers/requirements.txt
beautifulsoup4==4.10.0
pdfminer.six==20211012
requests==2.27.1
